# iteh code/eventapi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EventHub API",
        "description": "API dokumentacija za EventHub aplikaciju",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Prikaz svih kategorija",
                "operationId": "5334779003e44eba3d2ed0af49466c35",
                "responses": {
                    "200": {
                        "description": "Uspešno vraćena lista kategorija"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Kreiranje nove kategorije",
                "operationId": "35b7c7c4ff6b6d0eaa747a2c5688a517",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Konferencije"
                                    },
                                    "opis": {
                                        "type": "string",
                                        "example": "Događaji tipa konferencija"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kategorija uspešno kreirana"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    },
                    "403": {
                        "description": "Zabranjen pristup"
                    },
                    "422": {
                        "description": "Greška validacije"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Prikaz jedne kategorije",
                "operationId": "210ea397cf587928bc8458be518de086",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID kategorije (idCategory)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno vraćena kategorija"
                    },
                    "404": {
                        "description": "Kategorija nije pronađena"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Izmena postojeće kategorije",
                "operationId": "d4cd2fd7bf8555bcbb42ac55fa3b6f98",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID kategorije (idCategory)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ažurirane konferencije"
                                    },
                                    "opis": {
                                        "type": "string",
                                        "example": "Ažurirani opis kategorije"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kategorija uspešno izmenjena"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    },
                    "403": {
                        "description": "Zabranjen pristup"
                    },
                    "404": {
                        "description": "Kategorija nije pronađena"
                    },
                    "422": {
                        "description": "Greška validacije"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Brisanje kategorije",
                "operationId": "a77a46a20b44d8426a957fc28cc1e2ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID kategorije (idCategory)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kategorija uspešno obrisana"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    },
                    "403": {
                        "description": "Zabranjen pristup"
                    },
                    "404": {
                        "description": "Kategorija nije pronađena"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}